feat: Improved Schedule [ci]
This provides a more advanced schedule implementation along the lines of the features in ZIO.

Its simpler than ZIO, a schedule is just a IEnumerable<PositiveDuration>. Where a PositiveDuration is a wrapper type that does what is says on the tin.

The only additional type added was a transformer, that is just a Func<Schedule,Schedule>.

This provides most of what can be done with ZIO and the power to build stateful loops out of IEnumerable and build any sort of schedule that does not depend on the result and environment of the effect.

I have also pulled in some common constructors from Polly.