feat: Bake cancellation token into AffPure (#1057)
The current version of Aff<T> cannot be cancelled.

This simply bakes a cancellation token into the thunk. This makes an Aff<T> very similar to Async in F#, a cold task with automatic cancellation support.

BREAKING CHANGE: Fork now returns an Eff that can be used to cancel and rejoin the task, the same as Aff<RT,T>