Fix for Issue #1177 = Effectful producer
Removed `enumerate`, `enumerate2`, `observe`, and `observe2`.  Replaced with `yieldAll`.

Enumeration was playing a dual role, this was too confusing.  The expected behaviour of yielding all the items is now the only behaviour.

The behaviour that looped the remainder of the expression how now been removed.  This may be re-added later, but I don't think it's necessary.  Also, Transducers will have this capability in a future version (which will eventually be the backing monad for Pipes)

https://github.com/louthy/language-ext/issues/1177