feat: Support for Random IO [ci]
This implements an environment trait to support injectable random IO.

This can be extended to support new performance improvements in net6+ around generating fractional primitives.

This does not re-invent anything, it just wraps the default netcore Random class and adds enough functions on top to be useful. Inspiration from the Haskell and Elm Random modules was taken, but no additional type-class(es) was added, this can be looked into later, if the flexibility offered here is not enough.

Another great random module is the rust one, that was very powerful but overkill for what most C# dev needs.

Open to any and all comments and suggestions for improvement.