Attempt to fix C#14 build errors
As mentioned in #1501, there are two build errors introduced when trying to build with C#14 installed.
This PR resolves these issues with an explicit call to `.AsEnumerable()`. Alternative fix could be just to do a cast to `IEnumerable<T>`.

> 2>D:\git..\LanguageExt.Core\Immutable Collections\Arr\Arr.Module.cs(110,24,110,25): error CS0023: Operator '.' cannot be applied to operand of type 'void'
> 2>D:\git..\LanguageExt.Core\Immutable Collections\Arr\Arr.cs(605,29,605,30): error CS0023: Operator '.' cannot be applied to operand of type 'void'

Tested both on C#14 as well as with an explicit `LangVersion 13.0`, this compiles fine on both.